split testing to a test file
