+++ internal/fixture/state_test.go
@@ -1,0 +1,67 @@
+/*
+Copyright 2018 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package fixture
+
+import (
+	"bytes"
+	"fmt"
+	"testing"
+
+	"sigs.k8s.io/structured-merge-diff/typed"
+)
+
+func TestFixTabs(t *testing.T) {
+	cases := []struct {
+		in, out     typed.YAMLObject
+		shouldPanic bool
+	}{{
+		in:  "\t\ta\n\t\t\tb\n",
+		out: "a\n  b\n",
+	}, {
+		in:  "\n\t\ta\n\t\t\tb\n",
+		out: "a\n  b\n",
+	}, {
+		in:  "\n\t\ta\n\t\t\tb\n\t",
+		out: "a\n  b\n",
+	}, {
+		in:          "\t\ta\n\t\t  b\n",
+		shouldPanic: true,
+	}, {
+		in:          "\t\ta\n\tb\n",
+		shouldPanic: true,
+	}}
+
+	for i := range cases {
+		tt := cases[i]
+		t.Run(fmt.Sprintf("%v - %v", i, []byte(tt.in)), func(t *testing.T) {
+			if tt.shouldPanic {
+				defer func() {
+					if x := recover(); x == nil {
+						t.Errorf("expected a panic, but didn't get one")
+					}
+				}()
+			}
+			got := FixTabsOrDie(tt.in)
+			if e, a := tt.out, got; e != a {
+				t.Errorf("got %v", []byte(a))
+			}
+			if bytes.Contains([]byte(got), []byte{'\t'}) {
+				t.Error("contained a tab")
+			}
+		})
+	}
+}
